Remove vendored dependencies that actually are already in Debian
Despite multiple checks and many people reviewing, only the FTP Master
discovered that some of the dependencies vendored were in vain as they
actually already exist in Debian.

Only 'go-billy/v5' is a direct dependency of 'usql', so a later commit
may also remove the dependencies on 'google/goterm' and
'soniakeys/quant'.